Added in a function and statement for the crash avoidance
Added a function and statements that determine if any two different
agents will be able to even activate their crash avoidance. Only if the
statements are true will the two agents incremented projected
telemetries be compared. This was done to reduce run time but further
testing will need to be done.