WIP Optimizing Simulator
Working on optimizing the simulator so that the program runs faster. I
have been going through and and using the profiler tool finding the
functions were run time is slow and reducing the amount it is called
and allocating memory. So far I have been able to reduce the run time
from 33 second per generation down to roughly 2 given the current
parameters.